% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Check_Valid_Report.R
\name{Check_Valid_Report}
\alias{Check_Valid_Report}
\title{Check the Validity of the Input Poshmark Sales Activity File}
\usage{
Check_Valid_Report(t_datapath)
}
\arguments{
\item{t_datapath}{The file to be read.}
}
\value{
valid_report, valid_user Logical values in array.
}
\description{
This function Checks the Validity of the Input Poshmark Sales Activity File.
}
\examples{

}
