% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clean_Summary_Table.R
\name{Clean_Summary_Table}
\alias{Clean_Summary_Table}
\title{Clean Summary Tables for Dashboards}
\usage{
Clean_Summary_Table(t_table)
}
\arguments{
\item{t_table}{The table to be cleaned.}
}
\value{
t_table The cleaned table.
}
\description{
This function Cleans Summary Tables for Dashboards.
}
