% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fix_Bin_Number.R
\name{Fix_Bin_Number}
\alias{Fix_Bin_Number}
\title{Fixes Inventory Bin Numbers}
\usage{
Fix_Bin_Number(inv_table)
}
\arguments{
\item{inv_table}{The Inventory Table for the user.}
}
\value{
Updated inv_table as data frame.
}
\description{
This function fixes inventory bin numbers from old format to new. May be removed when no longer needed.
}
\examples{

}
