% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fix_Blank_SKU.R
\name{Fix_Blank_SKU}
\alias{Fix_Blank_SKU}
\title{Fixes Blank SKU in Poshmark Sales Activity Tables}
\usage{
Fix_Blank_SKU(sales_activity)
}
\arguments{
\item{sales_activity}{Poshmark Sales Activity Table}
}
\value{
sales_activity Updated Poshmark Sales Activity Table
}
\description{
This function Fixes Blank SKUs in Poshmark Sales Activity Tables.
}
\examples{

}
