% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fix_Partial_SKU.R
\name{Fix_Partial_SKU}
\alias{Fix_Partial_SKU}
\title{Fixes Partial SKU in Poshmark Sales Activity Tables}
\usage{
Fix_Partial_SKU(sales_activity)
}
\arguments{
\item{sales_activity}{Poshmark Sales Activity Table}
}
\value{
sales_activity Updated Poshmark Sales Activity Table
}
\description{
This function Fixes Partial SKUs in Poshmark Sales Activity Tables.
}
\details{
Requires: stringr
}
\examples{

}
