% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Format_Alert_Text_Confirm.R
\name{Format_Alert_Text_Confirm}
\alias{Format_Alert_Text_Confirm}
\title{Format Poshmark Confirmation Panel}
\usage{
Format_Alert_Text_Confirm(sales_activity)
}
\arguments{
\item{sales_activity}{Poshmark Sales Activity Table}
}
\value{
alert_text HTML Formatted Panel
}
\description{
This function formats the Poshmark Confirmation Panel.
}
\details{
Requires: shiny
}
\examples{

}
