% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Format_Alert_Text_Upload.R
\name{Format_Alert_Text_Upload}
\alias{Format_Alert_Text_Upload}
\title{Format Poshmark Upload Panel}
\usage{
Format_Alert_Text_Upload(sales_activity)
}
\arguments{
\item{sales_activity}{Poshmark Sales Activity Table}
}
\value{
alert_text HTML Formatted Panel
}
\description{
This function formats the Poshmark Upload Panel.
}
\details{
Requires: shiny
}
\examples{

}
