% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get_Changed_Data.R
\name{Get_Changed_Data}
\alias{Get_Changed_Data}
\title{Get Changed Data from Two Data Frames}
\usage{
Get_Changed_Data(source_df, target_df)
}
\arguments{
\item{source_df}{The Source Data Frame.}

\item{target_df}{The Target Data Frame.}
}
\value{
man_sku The next SKU for inventory items.
}
\description{
This function determines the changed data between two data frames.
}
\details{
Requires:
}
\examples{

}
