% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get_Changed_Goals.R
\name{Get_Changed_Goals}
\alias{Get_Changed_Goals}
\title{Get Changed User Goals}
\usage{
Get_Changed_Goals(source_df, target_df)
}
\arguments{
\item{source_df}{Source data drame}

\item{target_df}{Target data frame}
}
\value{
changed_tbl Changed table as data frame
}
\description{
This function finds changed user goals.
}
\details{
Requires:
}
\examples{

}
