% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get_Next_SKU.R
\name{Get_Next_SKU}
\alias{Get_Next_SKU}
\title{Get Next SKU for Inventory}
\usage{
Get_Next_SKU()
}
\value{
man_sku The next SKU for inventory items.
}
\description{
This function generates the next SKU for Inventory Table.
}
\details{
Requires: dplyr
}
\examples{

}
