% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GT_boxProfileT_Vals.R
\name{GT_boxProfileT_Vals}
\alias{GT_boxProfileT_Vals}
\title{Create Goal Tracker BoxProfileT Values}
\usage{
GT_boxProfileT_Vals(t_actual, t_goal, t_type)
}
\arguments{
\item{t_actual}{Actual values.}

\item{t_goal}{User Goal values.}
}
\value{
Formatted array of values for BoxProfileT
}
\description{
This function Creates Goal Tracker BoxProfileT Values.
}
\details{
Requires:
}
\examples{

}
