% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GT_InfoBoxT_Vals.R
\name{GT_InfoBoxT_Vals}
\alias{GT_InfoBoxT_Vals}
\title{Create Goal Tracker InfoBoxT Values}
\usage{
GT_InfoBoxT_Vals(t_actual, t_goal, t_type)
}
\arguments{
\item{t_actual}{Actual values.}

\item{t_goal}{User Goal values.}
}
\value{
Formatted array of values for InfoBoxT
}
\description{
This function Creates Goal Tracker InfoBoxT Values.
}
\details{
Requires:
}
\examples{

}
