% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Load_Add_Inv_Rec.R
\name{Load_Add_Inv_Rec}
\alias{Load_Add_Inv_Rec}
\title{Create Clean Inventory Record}
\usage{
Load_Add_Inv_Rec()
}
\value{
c_inv_record The clean inventory record.
}
\description{
This function creates a clean inventory record for add.
}
\details{
Requires:
}
\examples{

}
