% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Process_Valid_Report.R
\name{Process_Valid_Report}
\alias{Process_Valid_Report}
\title{Process Poshmark Sales Activity File}
\usage{
Process_Valid_Report(t_datapath)
}
\arguments{
\item{t_datapath}{The file to be read}
}
\value{
sales_activity The Poshmark Sales Activity table as data frame.
}
\description{
This function Processes the Poshmark Sales Activity File.
}
\details{
Requires: utils, dplyr, lubridate
}
\examples{

}
