% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Secure_Check.R
\name{Secure_Check}
\alias{Secure_Check}
\title{Perform Security Check on Tab Access}
\usage{
Secure_Check(input_tab, tab_name, user_auth)
}
\arguments{
\item{input_tab}{Current Tab}

\item{tab_name}{Target Tab Name}

\item{user_auth}{Authorized User}
}
\value{
auth_valid Logical return to allow/deny access.
}
\description{
This function performs a security check for the tab to be accessed.
}
\details{
Requires:
}
\examples{

}
