% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sort_Inv_Table.R
\name{Sort_Inv_Table}
\alias{Sort_Inv_Table}
\title{Sorts the Inventory Table}
\usage{
Sort_Inv_Table(inv_table)
}
\arguments{
\item{inv_table}{Inventory Table to Sort}
}
\value{
inv_table Sorted Inventory Table
}
\description{
This function Sorts the Inventory Table by Year, Month, Counter.
}
