% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Submit_Sales_Activity_DB.R
\name{Submit_Sales_Activity_DB}
\alias{Submit_Sales_Activity_DB}
\title{Submit Poshmark Sales Activity to Database}
\usage{
Submit_Sales_Activity_DB(sales_activity)
}
\arguments{
\item{sales_activity}{Poshmark Sales Activity Report}
}
\value{
sales_activity Poshmark Sales Activity Report
}
\description{
This function Submits the Poshmark Sales Activity to Database.
}
\details{
Requires: dplyr
}
\examples{

}
