% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Toast_Message.R
\name{Toast_Message}
\alias{Toast_Message}
\title{Generate Toast Message}
\usage{
Toast_Message(tr_message, tr_type)
}
\arguments{
\item{tr_message}{The message for the Toast output.}

\item{tr_type}{The type of Toast message to generate.}
}
\value{
A Toast message at the bottom right of the page.
}
\description{
This function generates a toast message.
}
\details{
Requires: shinytoastr
}
\examples{

}
