% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Update_Inventory_PM.R
\name{Update_Inventory_PM}
\alias{Update_Inventory_PM}
\title{Update Inventory Table with Poshmark Sales Activity Table}
\usage{
Update_Inventory_PM(d_sales_activity)
}
\arguments{
\item{d_sales_activity}{Poshmark Sales Activity Report}
}
\value{
Updates Inventory Table
}
\description{
This function Updates the Inventory Table with the Poshmark Sales Activity Table.
}
\details{
Requires:
}
\examples{

}
