% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Update_Quantity_Add.R
\name{Update_Quantity_Add}
\alias{Update_Quantity_Add}
\title{Update the Inventory Items to be Added}
\usage{
Update_Quantity_Add(inv_record, item_qty)
}
\arguments{
\item{inv_record}{The Inventory Record to be added.}

\item{item_qty}{The number of items to be added.}
}
\value{
inv_record The updated data frame with all inventory records to be added.
}
\description{
This function creates additional records for inventory add based on the item quantity.
}
\details{
Requires:
}
\examples{

}
