% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/User_Search_ID.R
\name{User_Search_ID}
\alias{User_Search_ID}
\title{Assigns the User ID Based on Role/Proxy Permissions}
\usage{
User_Search_ID()
}
\value{
Inventory Table as a data frame.
}
\description{
This function Assigns the User ID Based on Role/Proxy Permissions.
}
\details{
Requires: dplyr
}
